requirements.txt
chromadb==0.4.15
datasets==4.4.2
fastapi==0.128.0
numpy==1.24.3
pillow==12.1.0
pydantic==2.12.5
pytest==9.0.2
sentence_transformers==5.2.0
starlette==0.50.0
torch==2.9.1
uvicorn==0.40.0
requests==2.32.5
pydantic==2.12.5
huggingface-hub==0.36.0
gunicorn==21.2.0
datasets==4.4.2
pytest==9.0.2